testing

+++ sources/AACodonMutSelProfileProcess.cpp
@@ -131,7 +131,7 @@ double AACodonMutSelProfileProcess::LogCodonProfilePrior()	{
 void AACodonMutSelProfileProcess::SampleCodonProfile()	{
 	for (int i=0; i<statespace->GetNstate(); i++)  {
 		//codonprofile[i] = rnd::GetRandom().sExpo();
-		codonprofile[i] = 1.0/statespace->GetDegeneracy(i);
+		codonprofile[i] = 1.0/ (double) statespace->GetDegeneracy(i);
 	}
 }
